+++ clipboard-exe/Properties/AssemblyInfo.cs
@@ -10,8 +10,8 @@ using System.Windows;
 
 [assembly: AssemblyTitle("Clipboard")]
 [assembly: AssemblyProduct("Clipboard")]
-[assembly: AssemblyVersion("0.7.0.0")]
-[assembly: AssemblyFileVersion("0.7.0.0")]
-[assembly: AssemblyInformationalVersion("0.7.0")]
+[assembly: AssemblyVersion("0.7.3.0")]
+[assembly: AssemblyFileVersion("0.7.3.0")]
+[assembly: AssemblyInformationalVersion("0.7.3")]
 [assembly: TargetPlatform("Windows")]
 [assembly: ThemeInfo(ResourceDictionaryLocation.None, ResourceDictionaryLocation.SourceAssembly)]